Reorganize psfs subpackage and remove psfs_utils
