+++ scopesim/effects/psfs/__init__.py
@@ -1,4 +1,7 @@
 # -*- coding: utf-8 -*-
+from ...utils import get_logger
+logger = get_logger(__name__)
+
 from .psf_base import PSF, PoorMansFOV
 from .analytical import (Vibration, NonCommonPathAberration, SeeingPSF,
                          GaussianDiffractionPSF)
